Work in progress on tar enemy
Started working on his behavior. He is faster and has a lingering, spray like weapon that drips tar. Working on a random number system to give him a bit of a spray in his attack. In addition to this, would like to include a slight slowdown function from coming into contact with the tar.